subRoute()+ utilityBetweenStations()

+++ dotNet5781_3169_8515/dotNet5781_02_3169_8515/busLineStation.cs
@@ -30,7 +30,7 @@ namespace dotNet5781_02_3169_8515
             distance = -1;
             driveTime= new TimeSpan();
         }
-        internal busLineStation(string id, int lat, int lon,int distance, TimeSpan time, string address= "") :base(id, lat, lon, address)
+        internal busLineStation(string id, double lat, double lon,int distance, TimeSpan time, string address= "") :base(id, lat, lon, address)
         {
             this.distance = distance;
             this.driveTime = time;
